pyproject.toml
[project]
name = "langchain-course-campusx"
version = "0.1.0"
description = "Practice files of langchain course by campusx on youtube."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "bs4>=0.0.2",
 "faiss-cpu>=1.13.2",
 "grandalf>=0.8",
 "huggingface-hub>=0.36.1",
 "ipykernel>=7.2.0",
 "langchain>=1.2.8",
 "langchain-chroma>=1.1.0",
 "langchain-classic>=1.0.1",
 "langchain-community>=0.4.1",
 "langchain-core>=1.2.9",
 "langchain-huggingface>=1.2.0",
 "langchain-ollama>=1.0.1",
 "langchain-pinecone>=0.2.13",
 "langchain-text-splitters>=1.1.0",
 "pydantic[email]>=2.12.5",
 "pypdf>=6.7.0",
 "python-dotenv>=1.2.1",
 "pytube>=15.0.0",
 "scikit-learn>=1.8.0",
 "sentence-transformers>=5.2.2",
 "streamlit>=1.54.0",
 "tiktoken>=0.12.0",
 "wikipedia>=1.4.0",
 "youtube-transcript-api>=1.2.4",
]
